Updated openstack/openstack Project: openstack/cinder  16eaa4d5e19d974997161679592e428d58f61713
Scality SOFS: enhance how the remoteFS mount is detected

Previously we relied on a special "sys" dir being present at the root
of the mount point. This was weak. We now read the mount table and look
for a fuse-based FS mounted at the correct location.

Also, for the unit test, we slowly try to use mock instead of mox.

Change-Id: I055858e66db09fe7c5b1688c9fa05077948f342d