Updated openstack/openstack Project: openstack/sahara  619bba6d97195e83525ca9844276c14b3586fb3d
Removed dependency on Spark plugin in edp code

The EDP Spark engine was importing a config helper from the Spark
plugin.
The helper was moved to common plugin utils and now is imported from
there by both the plugin and the engine.

This is the part of sahara and plugins split.

Partially-implements bp: move-plugins-to-separate-repo
Change-Id: Ie84cc163a09bf1e7b58fcdb08e0647a85492593b