Updated openstack/openstack Project: openstack/ironic  3a46559c8e03997237e51454cda0a9e63f3c42b2
Allow vendor methods to serve static files

This patch adds a new parameter called 'attach' to the @passthru and
@driver_passthru decorators. If attach is True Ironic API will attach
the return value from the vendor method to the response object instead
of returning it in the response body (as it was before), this parameter
defaults to False.

Although this patch is a plumbing work for the blueprint
ipxe-dynamic-config it can be used separately from that. With this
extension any new vendor passthru method is able to serve static files
from the Ironic API.

Related-Blueprint: ipxe-dynamic-config
Change-Id: If2537efa35667aedd72581b4ed064992b00f9c18