Updated openstack/openstack Project: openstack/nova  39c4cc5eead59abf5241e010eaa701adb90e46d2
Avoid resizing disk if the disk size doesn't change

Sometimes we only want to resize the cpu and ram of the instance,
but nova still resizes and convert disk type even if the disk size
remains the same with the original, which will cost lots of time.

So we need to avoid resizing disk if the disk size doesn't change.

Change-Id: Iea1dd794ae8f9314a16bdf7143e0a639921a4b3c
Closes-Bug: #1454165