Updated openstack/openstack Project: openstack-infra/puppet-askbot  a07962ba554d0a56b6d60bc31da2ef449c731baf
Add Gemfile and puppet 4 checks

In anticipation of puppet 4, start trying to deal with puppet 4 things
that can be helpfully predicted by puppet lint plugins. This patch also
corrects lint errors caught by the puppet-lint-absolute_classname-check
gem.

We disable the puppet-lint-empty_string-check gem to be corrected by a
follow-up patch.

We disable the disable_arrow_alignment check in the Rakefile.
Previously, the system version of puppet-lint that was being run by
'rake lint' was too old to catch this. When using 'bundle exec rake
lint' the gem is new enough to catch this. Whether we want to actually
fix these lint errors is a different discussion, so this patch disables
it for now.

Change-Id: I2b2c8511367b2e97c30b7681125d9063b3e78a86