Updated openstack/openstack Project: openstack/nova  09a91292183f5e461fa8b2826f739f1a0fa608e5
compute: rename ResizeClaim to MoveClaim

We will introduce claiming for live-migrate and evacuate operations, and
the claiming logic is really the same for all of those (we will create
different Migration object though, but this is unrelated to claims).

Make sure that the Claim subclass name reflects this fact.

In addition rename a method on the resource tracker for dropping such
claims to 'drop_move_claim' to match.

Related-blueprint: migration-fix-resource-tracking
Change-Id: I9872f05174f9f79e755a8ac2585ce9a14760b26a