Updated openstack/openstack Project: openstack/nova  1b5b7e0c5de90758de93fce9a44996174564d4a2
virt: add get_device_name_for_instance to the base driver class

Just as libvirt needs to decide on the device names when booting an
instance (as opposed to them being assigned by the user), the same needs
to be done when attaching additional volumes to a running instance.

This patch exposes it on the ComputeDriver base class and makes sure
that it will be called by the compute manager (and that an appropriate
fallback will happen if the drivers do not care about overriding this),
as a pre-req for doing the work in the libvirt driver.

Partial-bug: 1452224
Related-bug: 1231874

Change-Id: I66693a8a6a632b65f930a58905336397ac6c9f29