Updated openstack/openstack Project: openstack/requirements  407e3b2d028a1311e74b18f64a57963369f1bf02
Make mock 1.1 the minimum version.

1.1 has numerous new features backported from cPython 3.5, and
crucially fixes Python 3.4 support allowing us to get rid of the six
hacks we've been using to use mock on 3.4.

funcsig comes into upper-constraints as a transitive dep, but as we
have no dep on it in our projects, its not added to
global-requirements.txt.

As we still have some 2.6 jobs around, I've pinned mock for 2.6 to
1.0.1, the last release of mock for 2.6.

Change-Id: Ibd2410e0153053b5121155474e99752256c7e4b8