Updated openstack/openstack Project: openstack-dev/devstack  45f0e3d0572f247c0487dd1df9cb15d39b7c0cf4
Use project instead of tenant in create_userrc

The create_userrc file is littered with references to a tenant. The tenant
concept has been deprecated long enough that we should use project instead.

I have not attempted to maintain compatibility with the --os-tenant-X flags
because I have not found reference to anyone using this script outside of
devstack.

Change-Id: I613f1bdc6673f0c4bfe29aaab7b514348a617a8c