Updated openstack/openstack Project: openstack/nova  4442979e71bb09ff9b71604c0dddd37c2bede04d
Fix for mock-1.1.0

This includes 2 classes of fixes (and 1 skip) for mock.

The first is the change in allowed values for assert_has_calls -
https://github.com/testing-cabal/mock/issues/263

The second is a yet unsolved bug around the use of autospec
https://github.com/testing-cabal/mock/issues/264

The skip is because something has changed with mock.open that is
causing the vhduils test to fail. We don't know why, but it's one test
to skip.

This also includes a brute force fix for tox -e docs, because pip is
no longer respecting the sphinx pin in test requirements. This has to
be landed with the other changes because they won't work without it.

Change-Id: Id835d080a1ada52cbd3f24dad9bab5eeb2f29a54
Partial-Bug: 1473401