Updated openstack/openstack Project: openstack/oslo.log  ee112220e630fd457038bcb5911eb3486c279dab
Do not report deprecations in subclasses

Upgrading to mock 1.1.0 somehow exposed the logic error in deprecated()
that was causing subclasses of deprecated exceptions to be reported as
deprecated when they should not be. The problem makes sense because we
were indiscriminately reporting the case in the constructor, and if the
subclass initialized the base class the deprecation reporting function
would be called. Change the logic to only report the deprecation if the
class being instantiated matches one of the deprecated classes.

Add a warning, and skip the test, for catching deprecated exceptions
under Python 3, since this feature does not always work there.

Also explicitly declare the versions of mock supported in the unit
tests, since we need different versions for python 2.6 vs. 2.7 and
later.

Change-Id: Ib60413a0e45901902382c3ae547be0e24df4e50c