Updated openstack/openstack Project: openstack/glance_store  ec025eb9e5502d9111b32dc5360a4bd5d9aff8eb
Remove usage of assert_called_once in mocks

This was never a real part of the mock api, but mock was happily mocking
it out and letting the tests pass. The new mock release made this
behaviour break, causing tests to fail on all patches. This patch
replaces instances of assert_called_once with an assert of the mock's
call_count.

Change-Id: I41ea0456acda86147ce3a7da0311e926782b4542
Closes-Bug: #1473369