Updated openstack/openstack Project: openstack/tempest-lib  71bc71a4d24bf1f7c46993e25c225bbbeace7ce3
Properly decode body before calling json.loads()

This commit fixes an issue when running the rest client and token
clients on python 3. httplib2 sets the type for the response body as
'bytes' when running on python 3, which requires that it be decoded
prior to running json.loads() on it. Additionally, in the v2 token
client a type check was done on the response body which was no longer
evaluating true because the type was no longer a 'str'. This was
fixed as part of the broader cleanup.

Change-Id: If4d496c4f10cec7d7050afc4b07f1f263de4c3e5