Updated openstack/openstack Project: openstack/heat-translator  1bdc0ad2a5d8ed28efa78662d4143e61865ed18a
Add app in the nodejs mongodb template

The current template does not make much sense as it's having nodejs and
mongodb deployed on two different instances but without any connection
between them. Add a nodejs app on the app server which uses mongodb database.
Nodejs custom type is not needed as it can use webserver type. Also, a better
test will be added with mapping to actual HOT template, not just resources, as
part of generator side work.

Partial-Bug: #1452545

Change-Id: I7f4ed72368bf7fcc89da6fbe9b322b8c6872e100