Updated openstack/openstack Project: openstack-dev/devstack  67011910d23957c12313ebff6bbdb3d234609615
Include domain variables in accrc

Include the user and project domain parameters in the generated user rc files.
This is fairly simplistic, if we were to follow the existing attitudes we
should iterate over the domains and add a new level of folders however this
would change the output location for files that may be depended upon.

Change-Id: I5e9e78406b11382751a591d91f711161bb98f47a