Updated openstack/openstack Project: openstack-infra/puppet-project_config  9f9049c323a7380dc0ff52a51fcd5dfa9a7a7f56
Add a common data directory for job workers

It's useful to be able to add data files on our workers as a common
reference, much in the same way job scripts are provided. This will
be used to keep a local copy of the bindep fallback package list so
that it doesn't require directly cloning the project-config repo
during job run-time.

Change-Id: I30a8be9167cc5c52faae59a35a284b3f37b93f44