Updated openstack/openstack Project: openstack/keystone  6881a76b3f2a5c95283a5ff93093b3f2e1adf3e9
Removed optional dependency support

This was sort of an anti-pattern anyway. You can't really depend on
something that's optional. If the need for something like this arises we
should look into established patterns like: template, decorator and
chain of responsibility. I believe that one or a combination of those
could have been used instead of the original implementation.

bp object-dependency-lifecycle

Change-Id: I65c834497d62325b29c95efe01d4be4b40ace37a