Updated openstack/openstack Project: openstack-infra/project-config  ae50d3408bf0c01f54716e8aba8d3ecd1f631549
Rename keystoneclient-saml2 to keystoneauth-saml2

Saml2 authentication plugins were initially depending on
python-keystoneclient. However with keystoneauth library to be released
soon we want to rename the repository to accurately reflect contents as
well as repository relations (it depends on keystoneauth not
keystoneclient). The 'python-' prefix is being dropped to be in line
with the non-prefixed "keystoneauth" repository.

Related governance change: https://review.openstack.org/#/c/190619/

Change-Id: I40de86d73694bdcf3177f6966274a103ac4aa2ad