Updated openstack/openstack Project: openstack-infra/puppet-asterisk  140508f8a0424e92ddee7c526acc6c470468e720
Add Gemfile and puppet 4 checks

In anticipation of puppet 4, start trying to deal with puppet 4 things
that can be helpfully predicted by puppet lint plugins. This patch also
corrects lint errors caught by the puppet-lint-absolute_classname-check
and puppet-lint-empty_string-check gems.

We disable the disable_arrow_alignment check in the Rakefile.
Previously, the system version of puppet-lint that was being run by
'rake lint' was too old to catch this. When using 'bundle exec rake
lint' the gem is new enough to catch this. Whether we want to actually
fix these lint errors is a different discussion, so this patch disables
the check for now.

Change-Id: Ib44f2d6ac941e6fed118a5ff78bddd3d0f0a40b5