Updated openstack/openstack Project: openstack-infra/project-config  f954e8ccd4d911a9314a588c5cf9d6fa9ae06d5b
Revert "Block installing of mock from wheels"

This reverts commit 71d10ea626f5a62d6b48133e19f469b19c9a7f14.

It was unnecessary once the offending wheel got removed from PyPI,
and this never worked anyway. It was preventing the ability to boot
new nodes, erroring thusly:

    ./configure_mirror.sh: line 25: /tmp/pip.conf: Permission denied

Deleting the images to which the change was applied returned
nodepool to being able to boot new nodes again.

Change-Id: I0841c6a5a26cf5be22e2d8fea861bdceb0393842