Updated openstack/openstack Project: openstack/python-troveclient  398bc8e1d7b31e414652305199b4c5986faf9acc
Error message on cluster-create is misleading

When you execute cluster-create without specifying an instance, the
error generated by Trove API service accurately reflects the error in
the API call. In the API, the instance information is an 'instances'
parameter.

However, on the command line, cluster-create accepts an argument that
is --instance. Therefore the error message coming from the Trove API
service (which reads ERROR: Validation error: cluster 'instances' is a
required property (HTTP 400) is misleading.

It is also not very meaningful for the client to knowingly send an
incomplete request to the server to be told that it is
incomplete. Instead, the client now just validates this and prints a
meaningful message for the user.

This does mean that the Trove client still has a --instance argument
while the API has an instances collection; both of which are
meaningful.

Closes-Bug: #1461604

Change-Id: I2a6b9c206c8676141b962e2d0602d77962c9e4dd