Updated openstack/openstack Project: openstack/openstack-manuals  af6464156b9f1f0c04e78fc9ed718140bf60eff2
Move services from projects, users, roles

Services documentation was in the "projects, users, and roles"
section, but there was already a service management section. Also,
the "projects, users, and roles" section was missing subsections
for the different resources.

The service management section was only missing some text for how
to delete a service.

Closes-Bug: #1446235
Change-Id: I5d9fe64e09b30e059aa384a00bcacb4e847dee79