Updated openstack/openstack Project: openstack/heat  5a4a738bcb55518731a3fe9bc3bce3e99db09588
Rename api unit test folder name

This is a preparation to split the openstack api v1 test cases. The
current test case file is too large for maintenance. This patch renames
the openstack folder to openstack_v1 so that in future we (hopefully)
don't have to move them around again.

Change-Id: Ie939e2d6022e688e02c5c946346b223e7b319136