Updated openstack/openstack Project: openstack/tempest  d32ca9fc70ad2dba38a060f2116c59db3e624fd9
Re-factor neutron client for 'list' methods

Currently neutron client is written to auto generate
methods on fly. This is very difficult to debug and
make code very difficult to read.

As neutron client will be moved to lib, it will be better to
have its interfaces more readable and consistent with other
service clients.

This commit re-factor it for 'list_*' methods.

Change-Id: Ia9513edd0ed5c080bd988cdf3e44e67774c698f9