Updated openstack/openstack Project: openstack/openstack-manuals  7d6fad01e79d7903ecf67e7e223f32154b45a61d
Use openstack rather than keystone

The keystone command is deprecated and should be not used in the
examples. The openstack command is the replacement.

The openstack command has handy options for extracting fields
so some of the commands were simplified.

The openstack command can use names when creating role
assignments and since this is more user-friendly the examples
were changed to use names rather than IDs.

Change-Id: Ic118284183001d10322cf357314672c5d98856a3